requirements.txt
absl-py==0.13.0
ansiwrap==0.8.4
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
Automat==20.2.0
Babel==2.9.1
backcall
backoff==1.11.1
beautifulsoup4==4.11.1
bleach==4.1.0
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.1
cssselect==1.1.0
cycler==0.11.0
debuglater==1.4.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.16.1
filelock==3.8.0
future==0.18.2
greenlet==1.1.3.post0
humanize==4.3.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.2.0
xgboost==1.7.1
incremental==22.10.0
ipdb==0.13.9
ipykernel==6.15.2
ipython==7.34.0
ipython-genutils==0.2.0
ipython-sql==0.4.1
ipywidgets==7.6.5
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.3.5
jupyter-console==6.4.3
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.1.2
jupyterlab_server==2.15.2
jupyterlab-widgets==1.0.0
jupytext==1.14.1
kiwisolver==1.4.4
lxml==4.9.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.1.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.0
mdurl==0.1.2
mediapipe==0.8.10
mistune==0.8.4
monotonic==1.6
nbclassic==0.3.5
nbclient
nbconvert==6.4.4
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.12
numpy==1.21.6
opencv-contrib-python==4.5.3.56
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
papermill==2.4.0
parsel==1.6.0
parso==0.8.3
pickleshare==0.7.5
pip==22.2.2
pkgutil_resolve_name==1.3.10
ploomber-core==0.0.6
ploomber-engine==0.0.10
ploomber-scaffold==0.3.1
ply==3.11
posthog==2.1.0
prettytable==0.7.2
prometheus-client==0.14.1
prompt-toolkit==3.0.20
Protego==0.2.1
protobuf==4.21.7
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.11.2
PyHive==0.6.5
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
#pywin32==302
#pywinpty==2.0.2
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
scipy==1.7.3
Scrapy==2.7.0
seaborn==0.10.0
Send2Trash==1.8.0
service-identity==21.1.0
setuptools==63.4.1
sip==6.6.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.42
sqlparse==0.4.2
tabulate==0.8.10
tenacity==8.0.1
terminado==0.13.1
testpath==0.6.0
textwrap3==0.9.2
tldextract==3.4.0
toml==0.10.2
tornado==6.2
traitlets
Twisted==22.8.0
#twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
urllib3
w3lib==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.8.0
zope.interface==5.5.0
ploomber==0.21.7
ploomber-core==0.0.6
ploomber-engine==0.0.10
ploomber-scaffold==0.3.1
dash==2.7.0
Flask==2.2.2 
Werkzeug==2.2.2 
dash-core-components==2.0.0 
dash-html-components==2.0.0 
dash-table==5.0.0 
itsdangerous==2.1.2
dash-bootstrap-components==1.2.1
scikit-learn==1.1.3
sklearn-evaluation
researchpy==0.3.5
patsy==0.5.3
statsmodels==0.13.5
Mako==1.2.4
alembic==1.9.2
cloudpickle==2.2.0 
databricks-cli==0.17.4
docker==6.0.1 
gitdb==4.0.10 
gitpython==3.1.30 
wandb==0.13.9
tensorflow==2.11.0